feat: fraudsight

+++ src/v1/services/Transaction.ts
@@ -11,6 +11,10 @@ import { blacklistedBins } from '@/helpers/constants';
 import { utils } from '@/helpers/utils';
 import { WorldpayError } from './Error';
 import { WorldpayVerifiedTokenRequest } from '../types/worldpay/verifiedToken';
+import {
+  WorldpayFraudOutcomeTypes,
+  WorldpayRiskAssessmentRequest,
+} from '../types/worldpay/fraudSight';
 
 export class TransactionService {
   private static instance: TransactionService;
@@ -28,31 +32,67 @@ export class TransactionService {
     return TransactionService.instance;
   }
 
+  private async handleWorldpayTransaction(
+    details: TransactionProcessProcessorDetails
+  ): Promise<WorldpayAuthorizePaymentResponse | Error | WorldpayError> {
+    try {
+      const {
+        verifiedTokenPayload,
+        authorizePaymentPayload,
+        fraudSightPayload,
+      } = details;
+      const { _embedded } =
+        await this.worldpayService.createVerifiedToken(verifiedTokenPayload);
+      const bin = _embedded.token.paymentInstrument.bin;
+
+      if (blacklistedBins.includes(bin)) {
+        throw new Error(`BIN ${bin} is blacklisted.`);
+      }
+
+      const tokenUrl = _embedded.token.tokenPaymentInstrument.href;
+      this.updatePaymentPayloads(
+        authorizePaymentPayload,
+        fraudSightPayload,
+        tokenUrl
+      );
+
+      authorizePaymentPayload.instruction.paymentInstrument.tokenHref =
+        tokenUrl;
+      fraudSightPayload.instruction.paymentInstrument.href = tokenUrl;
+      console.log(fraudSightPayload);
+
+      const riskAssessment =
+        await this.worldpayService.getRiskAssessment(fraudSightPayload);
+      if (riskAssessment.outcome === WorldpayFraudOutcomeTypes.HIGH_RISK) {
+        throw new Error('Transaction is high risk.');
+      }
+
+      // TODO: Link the FraudSight assessment
+      await this.worldpayService.authorizePayment(authorizePaymentPayload!);
+      return await this.worldpayService.deleteVerifiedToken(
+        utils.extractTokenFromUrl(tokenUrl)
+      );
+    } catch (error) {
+      throw error;
+    }
+  }
+
+  private updatePaymentPayloads(
+    authorizePayload: WorldpayAuthorizePaymentRequest,
+    fraudPayload: WorldpayRiskAssessmentRequest,
+    tokenUrl: string
+  ) {
+    authorizePayload.instruction.paymentInstrument.tokenHref = tokenUrl;
+    fraudPayload.instruction.paymentInstrument.href = tokenUrl;
+  }
+
   async processTransaction(
     processor: TransactionProcessor,
-    processorDetails: TransactionProcessProcessorDetails
+    worldpayProcessorDetails: TransactionProcessProcessorDetails
   ): Promise<WorldpayAuthorizePaymentResponse | Error | WorldpayError> {
     switch (processor) {
       case TransactionProcessor.WORLDPAY:
-        try {
-          const { verifiedTokenPayload, authorizePaymentPayload } =
-            processorDetails.worldpayProcessorDetails;
-          const { _embedded } = await this.worldpayService.createVerifiedToken(
-            verifiedTokenPayload!
-          );
-          if (blacklistedBins.includes(_embedded.token.paymentInstrument.bin)) {
-            // TODO
-          }
-          const tokenUrl = _embedded.token.tokenPaymentInstrument.href;
-          authorizePaymentPayload!.instruction.paymentInstrument.tokenHref =
-            tokenUrl;
-          await this.worldpayService.authorizePayment(authorizePaymentPayload!);
-          return await this.worldpayService.deleteVerifiedToken(
-            utils.extractTokenFromUrl(tokenUrl)
-          );
-        } catch (error) {
-          throw error;
-        }
+        return this.handleWorldpayTransaction(worldpayProcessorDetails);
       // Add more here
       default:
         throw new Error('Invalid payment processor');
